[REFACTOR] 디렉토리 변경으로 인한 import 문제 수정
