+++ WhyNComing/src/main/java/org/sparta/whyncoming/product/application/service/ProductService.java
@@ -1,9 +1,8 @@
 package org.sparta.whyncoming.product.application.service;
 
 import lombok.extern.slf4j.Slf4j;
-import org.sparta.whyncoming.common.response.ErrorCode;
+import org.sparta.whyncoming.common.exception.ErrorCode;
 import org.sparta.whyncoming.product.domain.entity.Category;
-import org.sparta.whyncoming.product.domain.entity.CategoryProduct;
 import org.sparta.whyncoming.product.domain.entity.Product;
 import org.sparta.whyncoming.product.domain.repository.CategoryRepository;
 import org.sparta.whyncoming.product.domain.repository.ProductRepository;
@@ -19,7 +18,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import static org.sparta.whyncoming.common.response.ErrorCode.*;
+import static org.sparta.whyncoming.common.exception.ErrorCode.INVALID_REQUEST;
+import static org.sparta.whyncoming.common.exception.ErrorCode.NOT_FOUND;
+
 
 @Slf4j
 @Service
@@ -97,7 +98,7 @@ public class ProductService {
     //TODO 반환값 String 하드코딩 대신 쓸 타입을 정해야 하고, 이미 삭제된 상품에 대한 예외처리가 필요함
     public String deleteProduct(UUID productId) {
         Product product = productRepository.findByProductId(productId)
-                .orElseThrow(() -> new IllegalArgumentException(String.valueOf(ErrorCode.NOT_FOUND)));
+                .orElseThrow(() -> new IllegalArgumentException(String.valueOf(NOT_FOUND)));
         product.delete();
 
         log.info("ProductService.deleteProduct() - after delete(), product: {}", product);
