+++ WhyNComing/src/main/java/org/sparta/whyncoming/user/presentation/controller/UserControllerV1.java
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 
+import org.sparta.whyncoming.common.exception.ErrorCode;
 import org.sparta.whyncoming.common.response.ApiResult;
-import org.sparta.whyncoming.common.response.ErrorCode;
 import org.sparta.whyncoming.common.response.ResponseUtil;
 import org.sparta.whyncoming.user.application.service.UserServiceV1;
